Move field manipulation from field planner to field provider to reduce code length of planner and let it completely be about UI, not model management.
