+++ field_friend/system.py
@@ -110,7 +110,7 @@ class System(rosys.persistence.PersistentModule):
         rosys.on_repeat(watch_robot, 1.0)
 
         self.path_provider = PathProvider()
-        self.field_provider = FieldProvider()
+        self.field_provider = FieldProvider(self.gnss)
         width = 0.64
         length = 0.78
         offset = 0.36
