+++ tests/test_field_provider.py
@@ -2,12 +2,14 @@ import json
 from pathlib import Path
 
 import pytest
+import rosys
 
 from field_friend.automations import FieldProvider
+from field_friend.navigation import GnssSimulation
 
 
 def test_loading_from_old_persistence():
-    field_provider = FieldProvider()
+    field_provider = FieldProvider(GnssSimulation(rosys.driving.Odometer(rosys.hardware.WheelsSimulation())))
     field_provider.restore(json.loads(Path('tests/old_field_provider_persistence.json').read_text()))
     assert len(field_provider.fields) == 3
     field = field_provider.fields[1]
